Refactor package structure for entities and enums; add Comment and Post entities
