+++ server/src/main/kotlin/br/com/webbicho/server/shared/entities/Post.kt
@@ -1,0 +1,20 @@
+package br.com.webbicho.server.shared.entities
+
+import jakarta.persistence.*
+
+@Entity(name = "posts")
+class Post(
+    @Column(columnDefinition = "TEXT", nullable = false)
+    var content: String,
+
+    @Column(columnDefinition = "TEXT", nullable = true)
+    var file: String? = null,
+
+    @OneToMany(
+        mappedBy = "post",
+        cascade = [CascadeType.ALL],
+        orphanRemoval = true,
+        fetch = FetchType.LAZY
+    )
+    var comments: MutableList<Comment> = mutableListOf()
+) : WbEntity()
