+++ server/src/main/kotlin/br/com/webbicho/server/shared/enums/EUserRole.kt
@@ -1,4 +1,4 @@
-package br.com.webbicho.server.shared.models.enums
+package br.com.webbicho.server.shared.enums
 
 enum class EUserRole {
     USER,
